requirements.txt
fastapi
uvicorn
transformers
pymongo
python-dotenv
gunicorn
torch
python-jose[cryptography]
passlib[bcrypt]
azure-storage-blob
python-multipart
pandas
sendgrid
email-validator
